requirements.txt
stransi
Pillow
